removed: ODK Aggregate bug workaround for Infinity, -Infinity and NaN
This workaround doesn't really solve much, because calculations (for now) always have type=string in XLSForms. The only use case for the ODK Aggregate bug that it works around is where a readonly decimal or integer input has a calculation attached to it, or a handcoded XForm that has type=integer, or type=decimal for a calculation. Too obscure to work around.

https://github.com/opendatakit/opendatakit/issues/1116