Convert tests from Jasmine to Mocha/Chai/Sinon
This change makes testing in enketo-core more consistent with testing in centro, enketo-transformer and enketo-express (and likely other Enketo projects, those are just the ones I'm most familiar with).

### Benefits/motivation

This should be a development experience improvement, bringing API/test framework feature parity and reducing the cognitive load of switching between projects.

### Risks

Quite a lot of the necessary changes were identified by project-wide search for relevant interface changes (e.g. `toEqual` -> `.to.equal`). However some of those changes were not 100% API-equivalent. For instance some changes of `toEqual` needed to be changed to `to.deep.equal` when comparing reference types.

While those changes were easy to identify (those tests didn't pass with the naive `.to.equal` replacement), the inverse (`.not.toEqual` -> `.not.to.deep.equal`) is not as easily identified, and may result in tests passing where they should not.

I've done my best with the tools available to manually verify that I addressed these cases, but it's certainly possible that there are some I've missed.

It's possible that there are other API inconsistencies of this sort lurking, though feels unlikely based on my reading of Jasmine and Chai docs.