Fix: use form language for number inputs (fall back to browser)
Also fixes and unifies changed language behavior, and re-formatting values when decimal characters change. Moves visibility fix to styles, as it's a core stylesheet and should be safe to change there. It's not yet clear what a safe but more specific selector would be, but it is clear that transformer doesn't set translation `lang` atttributes directly on question elements