Fix: focus first active input/widget on load, page navigation, user-added repeat (#969)
* Fix: focus first active input/widget on load, page navigation, user-added repeat

* Additional focus fixes

- Don't infinitely recurse on TOC navigation
- Don't infinitely recurse on navigation to first/last page
- Don't try to focus when no form control is found

Also includes a test for focus on newly relevant control. Worth noting that this does _not_ solve the same case when navigating by tab key in a non-pages version of the same form.

* Only scroll in `goToTarget` if necessary

In most cases the browser will scroll to the focused element automatically, but this will fallback to scrolling to the question/target if that doesn't occur.